Added Player and Projectile Classes and Changed Main
I added a class for the Player and the Projectile. The Player class actually seems to be working fine for now, the projectile class needs to be worked on as it doesn't work. I have also added more images to use as sprites which I have found online but they are referenced in the document.