I am about half way into fixing the shooting projectile
I tried expanding on the projectile class a little bit. I have changed the sprite for the player and created my own sprite for the fireball. I tried making it so that the projectiles get added to an array but it doesn't work. I will attempt to fix it later.